requirements-dev.txt
# 開発環境用の依存パッケージ
pytest>=6.0.0
pyyaml
tomli
tomli-w
